requirements.txt
mss==9.0.1
opencv-python==4.10.0.84
numpy==2.1.0
Pillow==10.4.0
pytesseract==0.3.13
